Move docker and entoend scenario class to core
